+++ pcs/constraint.py
@@ -189,6 +189,15 @@ def _validate_constraint_resource(cib_dom, resource_id):
         utils.err(resource_error)
 
 
+def _validate_resources_not_in_same_group(cib_dom, resource1, resource2):
+    if not utils.validate_resources_not_in_same_group(
+        cib_dom, resource1, resource2
+    ):
+        utils.err(
+            "Cannot create an order constraint for resources in the same group"
+        )
+
+
 # Syntax: colocation add [role] <src> with [role] <tgt> [score] [options]
 # possible commands:
 #        <src> with        <tgt> [score] [options]
@@ -477,6 +486,8 @@ def _order_add(resource1, resource2, options_list, modifiers):
     cib_dom = utils.get_cib_dom()
     _validate_constraint_resource(cib_dom, resource1)
     _validate_constraint_resource(cib_dom, resource2)
+
+    _validate_resources_not_in_same_group(cib_dom, resource1, resource2)
 
     order_options = []
     id_specified = False